+++ Changelog.md
@@ -1,17 +1,26 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 3.1.5
+Fixed an error with boss borders
+___
 ### 3.1.4
+Added more border size options\
 Added gradient to player and target castbars when non interruptable\
 Added gradient to player and target castbars when interrupted\
 Added gradient and custom texture to Target of Target of Target\
 Added minimap border options to adjust its size\
 Added pet border options to adjust its size\
+Added V3 statusbar textures by Blinkii\
+Updated the nameplate debuff glow function and added an option to customize the timer\
+Updated install, adding borders and making Light/Dark a toggle\
+Updated gradient chat command to be a toggle, typing it while gradient is active will disable gradient\
 Updated gradient function, splitting player and target castbars\
 Updated some shadows to not apply when borders are enabled since borders would hide them\
-Updated borders to apply settings when changing them in Eltruism without requiring a reload, but a reload is still recommended after configuring them\
+Updated borders function to update without a reload\
 Updated borders to check for ElvUI Action Bar Masks being loaded\
 Updated auto adjust borders function to apply when changing specs in retail if the profiles are both from Eltruism\
+Fixed target class statusbar textures options in nameplates, which was applying player class instead of target\
 Fixed an issue where auto adjusting borders would not load with Eltruism profiles\
 Fixed an issue where borders could sometimes not be centered correctly on Player/Target unitframes\
 Fixed player and target castbar borders on different types of cast attachment\